Hosted_Agents/contoso-support-agent/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b7
agent-framework

azure-identity
redis
python-dotenv
